Add trigger window support for "pid" and "title" fields
When trigger APIs return a window object, that object will usually include an "id" field. But sometimes, the "id" field will be missing and a "pid" or "title" field will be present instead. This change updates the window handling support to look for those fallback fields and process them correctly.